Pack Organisaatiot to single file
