--- services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/organisaatio/Domain.kt
@@ -1,9 +1,0 @@
-package fi.hel.haitaton.hanke.organisaatio
-
-data class Organisaatio(
-    var id: Int? = null,
-    var tunnus: String? = null,
-    var nimi: String? = null
-)
-
-class OrganisaatioNotFoundException() : RuntimeException("Cannot find organizations")
