--- services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/organisaatio/OrganisaatioService.kt
@@ -1,5 +1,0 @@
-package fi.hel.haitaton.hanke.organisaatio
-
-interface OrganisaatioService {
-    fun getOrganisaatiot(): Iterable<Organisaatio>
-}
