--- services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/organisaatio/OrganisaatioServiceImpl.kt
@@ -1,19 +1,0 @@
-package fi.hel.haitaton.hanke.organisaatio
-
-import org.springframework.beans.factory.annotation.Autowired
-
-class OrganisaatioServiceImpl(@Autowired val organisaatioRepository: OrganisaatioRepository) : OrganisaatioService {
-
-    override fun getOrganisaatiot(): Iterable<Organisaatio> {
-        val entities = organisaatioRepository.findAllByOrderByNimiAsc()
-        return entities.map { createdOrganisaatioDomainObjectFromEntity(it) }
-    }
-
-    private fun createdOrganisaatioDomainObjectFromEntity(organisaatioEntity: OrganisaatioEntity): Organisaatio {
-        return Organisaatio(
-            organisaatioEntity.id,
-            organisaatioEntity.organisaatioTunnus,
-            organisaatioEntity.nimi
-        )
-    }
-}
